requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
mcp>=1.6.0
fastmcp>=2.2.1
ncbi-datasets-pylib>=0.3.0 